pyproject.toml
[tool.poetry]
name = "epicevent"
version = "0.1.0"
description = ""
authors = ["Michael 56340359+SynysterRev@users.noreply.github.com"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<4.0.0"
sqlalchemy = ">=2.0.38,<3.0.0"
psycopg2 = ">=2.9.10,<3.0.0"
dotenv = ">=0.9.9,<0.10.0"
argon2-cffi = ">=23.1.0,<24.0.0"
pyjwt = ">=2.10.1,<3.0.0"
crypto = ">=1.4.1,<2.0.0"
click = ">=8.1.8,<9.0.0"
cryptography = "^44.0.2"
pytest = "^8.3.5"
pytest-mock = "^3.14.0"
black = "^25.1.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
epicevent = "cli:cli"